Uploading File Prototype + Function deleted/added
Most of changes is for adding file uploading function.
Just added some and modified with already existed photo upload form. +
Changed name of some codes but not done yet. This is not a clear code.
Deleted adding url of photo function cause since we integrated
Summernote editor and it has that function.
Added new routes for photo_work page.